webapp/requirements.txt
Django==2.1.3
psycopg2
psycopg2-binary
django-tinymce
Pillow